+++ src/resources/js/Components/Forms/FormforUpdateChecklist.jsx
@@ -1,42 +1,17 @@
-import React, { useState, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faClipboardCheck } from '@fortawesome/free-solid-svg-icons';
-import { fetchCheckLists, addCheckList, selectCheckListsByFlowNumber } from '../store/checklistSlice'; // Adjust the import path as needed
-import '../../css/CheckListModalContent.css';
+import { useFormforUpdateChecklist } from '../../Hooks/useFormforUpdateChecklist';
+import '../../../css/FormforUpdateChecklist.css';
 
-const CheckListModalContent = ({ workflowId, flowNumber, checkListsForFlowNumber }) => {
-    const [name, setName] = useState('');
-    const [error, setError] = useState(null);
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        dispatch(fetchCheckLists(workflowId)); // フェッチする必要があれば
-    }, [dispatch, workflowId]);
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        setError(null);
-
-        const checklistData = {
-            name,
-            flownumber_for_checklist: flowNumber, // flowNumberを使用
-        };
-
-        try {
-            await dispatch(addCheckList({ workflowId, checklist: checklistData })).unwrap();
-            setName(''); // 入力をクリア
-            dispatch(fetchCheckLists(workflowId)); // チェックリストを再取得
-        } catch (error) {
-            if (error.response && error.response.status === 422) {
-                console.error('Validation errors:', error.response.data.errors);
-                setError('Validation error occurred.');
-            } else {
-                console.error('Error adding Checklist:', error);
-                setError('An error occurred while adding the checklist.');
-            }
-        }
-    };
+const FormforUpdateChecklist = ({ flowNumber, checkListsForFlowNumber }) => {
+    const {
+        // Local State
+        name, setName, error,
+        // Global State
+        // Event Handler
+        handleSubmit,
+    } = useFormforUpdateChecklist(flowNumber);
 
     return (
         <div>
@@ -76,4 +51,4 @@ const CheckListModalContent = ({ workflowId, flowNumber, checkListsForFlowNumber
     );
 };
 
-export default CheckListModalContent;
+export default FormforUpdateChecklist;
