+++ src/resources/js/Components/ModalforDocumentSettings.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../css/ModalforAddCheckListForm.css';
+import '../../css/ModalofFormforAddFlowstep.css';
 import { closeDocumentSettingsModal } from '../store/modalSlice';
 import { useDispatch } from 'react-redux';
